Move types to types package (#95)
* Move types to types package

* Remove types dependency on vscode protocol

* Update README.md

* Update lsp.ts to export only TextDocument type

* Keep export of TextDocument class implementation in protocol

---------
