+++ runtimes/package.json
@@ -38,9 +38,7 @@
         "jose": "^5.2.3",
         "rxjs": "^7.8.1",
         "vscode-languageserver": "^9.0.1",
-        "vscode-languageserver-protocol": "^3.17.5",
-        "vscode-languageserver-textdocument": "^1.0.11",
-        "vscode-languageserver-types": "^3.17.5"
+        "vscode-languageserver-protocol": "^3.17.5"
     },
     "devDependencies": {
         "@types/mocha": "^10.0.1",
