+++ runtimes/protocol/chat.ts
@@ -15,10 +15,16 @@ import {
     TabChangeParams,
     TabRemoveParams,
     VoteParams,
-} from '@aws/language-server-runtimes-types'
-import { ProtocolNotificationType, ProtocolRequestType } from './lsp'
+    ProtocolNotificationType,
+    ProtocolRequestType,
+    ProgressToken,
+} from './lsp'
 
-export const chatRequestType = new ProtocolRequestType<ChatParams, ChatResult, ChatResult, void, void>(
+export interface ChatRequest extends ChatParams {
+    partialResultToken?: ProgressToken
+}
+
+export const chatRequestType = new ProtocolRequestType<ChatRequest, ChatResult, ChatResult, void, void>(
     'aws/chat/sendChatPrompt'
 )
