+++ types/package.json
@@ -14,6 +14,7 @@
     "author": "Amazon Web Services",
     "license": "Apache-2.0",
     "dependencies": {
-        "vscode-languageserver-protocol": "^3.17.5"
+        "vscode-languageserver-textdocument": "^1.0.11",
+        "vscode-languageserver-types": "^3.17.5"
     }
 }
